api-service/pyproject.toml
[tool.poetry]
name = "photocorrectserver"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.1"
gunicorn = "^22.0.0"
pydantic-settings = "^2.3.4"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
